mainSite/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
Django==4.1.1
gTTS==2.2.4
idna==3.4
imutils==0.5.4
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
PyAudio==0.2.12
pyngrok==5.1.0
pyparsing==3.0.9
pypiwin32==223
pytesseract==0.3.10
pyttsx3==2.90
pywin32==305
PyYAML==6.0
requests==2.28.1
six==1.16.0
SpeechRecognition==3.8.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
